webscrape_many()
When invoking webscrape_many() we can either pass in the product pages links as a text file or else as an array.

Example 1: Passing in a file
webscrape_many('products.txt')

Example 2: Passing in an array
webscrape_many(['https://www.amazon.com/Foo-bar/dp/B07CRG94G3', 'https://www.amazon.com/What-ever/dp/H6T5FWR57U'])

This is known as Multiple Dispatch or multimethods. It is a feature of some object-oriented programming languages in which a function or method can be dynamically dispatched based on the run time (dynamic) type of more than one of its arguments.

Python does not support multiple dispatch out of the box which is why we are using the package multipledispatch.

Read more: https://stackoverflow.com/questions/6434482/python-function-overloading